Project1/requirements.txt
tensorflow
fastapi 
uvicorn
Pillow
python-multipart
Jinja2

# absl-py==1.4.0
# annotated-types==0.5.0
# anyio==3.7.1
# astunparse==1.6.3
# cachetools==5.3.1
# certifi==2023.7.22
# charset-normalizer==3.2.0
# click==8.1.7
# colorama==0.4.6
# fastapi==0.103.1
# flatbuffers==23.5.26
# gast==0.4.0
# google-auth==2.22.0
# google-auth-oauthlib==1.0.0
# google-pasta==0.2.0
# grpcio==1.56.2
# h11==0.14.0
# h5py==3.9.0
# idna==3.4
# Jinja2==3.1.2
# keras==2.13.1
# libclang==16.0.6
# Markdown==3.4.4
# MarkupSafe==2.1.3
# numpy==1.24.3
# oauthlib==3.2.2
# opt-einsum==3.3.0
# packaging==23.1
# Pillow==10.0.0
# protobuf==4.23.4
# pyasn1==0.5.0
# pyasn1-modules==0.3.0
# pydantic==2.3.0
# pydantic_core==2.6.3
# python-multipart==0.0.6
# requests==2.31.0
# requests-oauthlib==1.3.1
# rsa==4.9
# six==1.16.0
# sniffio==1.3.0
# starlette==0.27.0
# tensorboard==2.13.0
# tensorboard-data-server==0.7.1
# tensorflow==2.13.0
# tensorflow-estimator==2.13.0
# tensorflow-intel==2.13.0
# tensorflow-io-gcs-filesystem==0.31.0
# termcolor==2.3.0
# typing_extensions==4.7.1
# urllib3==1.26.16
# uvicorn==0.23.2
# Werkzeug==2.3.6
# wrapt==1.15.0
